Create README.md
EnergyPlus version 8.9 had shown that the heat from the VGHE can heat the indoors to 15C at night provided the ground below the hydronic floor has warmed up after 5-6 years of comfortable indoor temperatures. As the experimental building in Melbourne Australia is new, this warm up period has not been met. and a wood/sawdust burner was used to circulate hotter water for the hydronic floor. To meet the daytime comfort temperature of 20C, simulations had used 30 evacuated tube solar collector to heat 2 cubic meter of water stored in an indoor tank to 50C by the end of summer. This storage serves as a thermal mass to buffer the cold season. In this project, a more affordable solution is to be simulated. Instead of evacuated tube solar collector, a parabolic trough is used to reflect and concentrate the weak winter solar radiation on to the floor/ground, which is insulated at its perimeter to serve as the thermal mass .